+++ opentelemetry-sdk/src/opentelemetry/sdk/environment_variables/__init__.py
@@ -51,7 +51,7 @@ OTEL_LOG_LEVEL = "OTEL_LOG_LEVEL"
 .. envvar:: OTEL_LOG_LEVEL
 
 The :envvar:`OTEL_LOG_LEVEL` environment variable sets the log level used by the SDK logger.
-OTLP exporters will log partial success responses when this is set to `info`, `debug`, or `verbose`.
+OTLP exporters will log partial success responses when this is set to `debug` or `verbose`.
 Default: "info"
 """